refactor: rename open_router/api.rs to provider.rs
Improves code organization by renaming the OpenRouter implementation file
from api.rs to provider.rs to be more consistent with other provider
implementations in the codebase. Updates the corresponding module
references.
